requirements.txt
selenium>=4.21.0
pillow>=10.0.0
numpy>=1.23.0
gymnasium>=0.29.1
stable-baselines3>=2.3.0
torch>=2.0.0
